JS-7541: app settings structure
