+++ src/ts/component/page/main/settings/data/index.tsx
@@ -25,13 +25,9 @@ const PageMainSettingsDataIndex = observer(class PageMainSettingsDataIndex exten
 	};
 
 	render () {
-		const { onPage } = this.props;
-		const { list } = this.state;
 		const { dataPath, spaceStorage } = S.Common;
 		const { localUsage } = spaceStorage;
 		const suffix = this.getSuffix();
-		const size = U.File.size(list.reduce((acc, item) => acc + item.size, 0));
-		const isLocal = U.Data.isLocalNetwork();
 
 		return (
 			<>
@@ -68,14 +64,6 @@ const PageMainSettingsDataIndex = observer(class PageMainSettingsDataIndex exten
 						</div>
 					</div>
 				</div>
-
-				{!isLocal ? (
-					<>
-						<Title className="sub" text={translate('popupSettingsDataManagementDeleteTitle')} />
-						<Label className="description" text={translate('popupSettingsDataManagementDeleteText')} />
-						<Button className="c36" onClick={() => onPage('delete')} color="red" text={translate('popupSettingsDataManagementDeleteButton')} />
-					</>
-				) : ''}
 			</>
 		);
 	};
