+++ src/ts/component/page/main/settings/phrase.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { observer } from 'mobx-react';
-import { Title, Label, Phrase, QR } from 'Component';
+import { Title, Label, Phrase, QR, Button } from 'Component';
 import { I, C, S, U, translate, analytics, Storage, Renderer } from 'Lib';
 
 interface State {
@@ -27,6 +27,8 @@ const PageMainSettingsPhrase = observer(class PageMainSettingsPhrase extends Rea
 
 	render () {
 		const { entropy, showCode } = this.state;
+		const { onPage } = this.props;
+		const isLocal = U.Data.isLocalNetwork();
 
 		return (
 			<div
@@ -53,6 +55,14 @@ const PageMainSettingsPhrase = observer(class PageMainSettingsPhrase extends Rea
 						<QR value={showCode ? entropy : translate('popupSettingsCodeStub')} />
 					</div>
 				</div>
+
+				{!isLocal ? (
+					<>
+						<Title className="sub" text={translate('popupSettingsDataManagementDeleteTitle')} />
+						<Label className="description" text={translate('popupSettingsDataManagementDeleteText')} />
+						<Button className="c36" onClick={() => onPage('delete')} color="red" text={translate('popupSettingsDataManagementDeleteButton')} />
+					</>
+				) : ''}
 			</div>
 		);
 	};
